Update logo crown to the modern version
During the last logo update iteration, the legacy
version is still used as the standard png image
for repositories.

This commit moves old versions to a legacy repository
and updates the "crown" version SVG and PNG files.

With this change, any place that references the old
PNG version will display correctly regardless of
using light or black themes, like the official GitHub
respository.

I also updated the README to refleck the context
changes of the updated logo images.